third_party/WebKit/LayoutTests/platform/win/external/wpt/IndexedDB/idbobjectstore_put8-expected.txt

This is a testharness.js-based test.
PASS IDBObjectStore.put() - object store has autoIncrement:true and the key path is an object attribute  
Harness: the test ran to completion.
